# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Prompt & Skill Registry API
  version: 0.2.0
  description: 'Contrato REST preliminar da versão Cloudflare: browser via Cloudflare Access + app RBAC; integração
    bearer Registry token com Access Service Auth na borda quando habilitado. Não é token Cloudflare API. Conforme
    docs/05-auth-and-policy.md.'
servers:
- url: http://localhost:5173
  description: Local Worker dev (exemplo)
- url: https://registry.example.com
  description: Placeholder, não é endpoint ativo; proteger com Cloudflare Access
security:
- BearerAuth: []
paths:
...
  /v1/audit-logs:
    get:
      operationId: list_audit_logs
      summary: Eventos de auditoria autorizados
      tags:
      - Audit
      parameters:
      - name: asset_id
        in: query
        required: false
        schema:
          type: string
      - name: cursor
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/executions:
    get:
      operationId: list_executions
      summary: Traces de execução/resolução
      tags:
      - Runtime
      parameters:
      - name: asset_id
        in: query
        required: false
        schema:
          type: string
      - name: cursor
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    post:
      operationId: report_execution
      summary: Aplicação reporta execução após resolve
      tags:
      - Runtime
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecutionReport'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Execution'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/webhooks:
    get:
      operationId: list_webhooks
      summary: Lista destinos de webhook
      tags:
      - Webhooks
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WebhookEndpoint'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    post:
      operationId: create_webhook
      summary: Cria destino HTTPS
      tags:
      - Webhooks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookEndpointCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookEndpoint'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/webhooks/{id}:
    patch:
      operationId: update_webhook
      summary: Atualiza destino HTTPS
      tags:
      - Webhooks
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookEndpointCreate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookEndpoint'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      operationId: disable_webhook
      summary: Desativa destino HTTPS
      tags:
      - Webhooks
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API token or user access token
  responses:
    BadRequest:
      description: Validação/entrada inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: Não autenticado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: Permissão insuficiente
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Recurso ausente ou invisível
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: Estado concorrente ou já existe
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    ApiError:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    AuditList:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            additionalProperties: true
        next_cursor:
          type: string
          nullable: true
    ExecutionReport:
      type: object
      required:
      - resolution_trace_id
      - status
      properties:
        resolution_trace_id:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - success
          - failure
        model:
          type: string
        latency_ms:
          type: integer
        input_hash:
          type: string
        output_hash:
          type: string
        cost_usd:
          type: number
    Execution:
      type: object
      additionalProperties: true
    ExecutionList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Execution'
        next_cursor:
          type: string
          nullable: true
    WebhookEndpointCreate:
      type: object
      required:
      - url
      properties:
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
    WebhookEndpoint:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
        enabled:
          type: boolean
        events:
          type: array
          items:
            type: string
...